fix(#2456): split pressure_drop_calculation_engine.py (1188 LOC) into three modules
Split the monolithic engine file into focused modules under a coordinating facade:
- _friction_factors.py (296 LOC): Colebrook, Swamee-Jain, Churchill, Haaland correlations
- _flow_calculations.py (577 LOC): flow properties, regime classification, pressure drop functions
- pressure_drop_calculation_engine.py (404 LOC): coordinator facade + PressureDropCalculationEngine class

All public names remain importable from the original module path (backward compat).
Adds TDD contract tests in tests/unit/upstream_drift_tools/test_pressure_drop_engine_split_2456.py.
